requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
openai==1.3.5
google-api-python-client==2.108.0
oauth2client==3.0.0
redis==5.0.1
feedparser==6.0.10
httpx==0.25.1
python-multipart==0.0.6
aiohttp==3.9.1
